methods/WBC_chall-main/requirements.txt
# WBCBench 2026 - White Blood Cell Classification Challenge
# Requirements for deep learning-based classification

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
albumentations>=1.3.0

# Data Science & ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress & Utilities
tqdm>=4.66.0

# Kaggle API
kaggle>=1.5.0

# Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Pretrained Models
timm>=0.9.0

# Additional utilities
pyyaml>=6.0
tensorboard>=2.14.0

# For handling imbalanced datasets
imbalanced-learn>=0.11.0

# For metrics and evaluation
scipy>=1.11.0

